# docs/swagger.yaml
basePath: /
definitions:
...
  api.ExtendedOrderResponse:
    properties:
      change_amount:
        type: number
      description:
        type: string
      orderInfo:
        $ref: '#/definitions/api.OrderResponse'
      products:
        items:
          $ref: '#/definitions/api.OrderProduct'
        type: array
      received_amount:
        type: number
      statuses:
        items:
          $ref: '#/definitions/api.OrderStatus'
        type: array
    type: object
...
  api.OrderClient:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      phone:
        type: string
    type: object
  api.OrderProduct:
    properties:
      article:
        type: string
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
    type: object
  api.OrderResponse:
    properties:
      client:
        $ref: '#/definitions/api.OrderClient'
      created_at:
        type: string
      id:
        type: integer
      status:
        type: string
      total_amount:
        type: number
    type: object
  api.OrderStatus:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  api.OrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/api.OrderResponse'
        type: array
    type: object
...
  github_com_phenirain_sso_internal_dto_auth.AuthRequest:
    properties:
      login:
        description: Логин пользователя
        example: user@example.com
        type: string
      password:
        description: Пароль пользователя
        example: P@ssw0rd!
        type: string
    type: object
  github_com_phenirain_sso_internal_dto_auth.AuthResponse:
    properties:
      access_token:
        description: Access Token для доступа к защищенным ресурсам
        type: string
      refresh_token:
        description: Refresh Token для обновления пары токенов
        type: string
    type: object
...
  github_com_phenirain_sso_internal_dto_response.Response-api_ExtendedOrderResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/api.ExtendedOrderResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
...
  github_com_phenirain_sso_internal_dto_response.Response-api_OrdersResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/api.OrdersResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
...
  github_com_phenirain_sso_internal_dto_response.Response-report_AverageOrderProcessingTimeResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/report.AverageOrderProcessingTimeResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
  github_com_phenirain_sso_internal_dto_response.Response-report_OrdersByTimeOfDayResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/report.OrdersByTimeOfDayResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
  github_com_phenirain_sso_internal_dto_response.Response-report_PurchasesByBrandsResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/report.PurchasesByBrandsResponse'
        description: Данные ответа
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
  github_com_phenirain_sso_internal_dto_response.Response-string:
    properties:
      data:
        description: Данные ответа
        type: string
      details:
        description: Детали ошибки
        type: string
      message:
        description: Сообщение (комментарий) об ошибке
        type: string
      success:
        description: Статус ответа
        type: boolean
    type: object
  manager.PaidOrderRequest:
    properties:
      order_id:
        type: integer
      paid_amount:
        type: number
    type: object
...
  report.AverageOrderProcessingTimeResponse:
    properties:
      units:
        items:
          $ref: '#/definitions/report.AverageOrderProcessingTimeResponse_Unit'
        type: array
    type: object
  report.AverageOrderProcessingTimeResponse_Unit:
    properties:
      avg_time_to_checkout_minutes:
        type: number
      avg_time_to_receive_minutes:
        type: number
    type: object
  report.OrdersByTimeOfDayResponse:
    properties:
      units:
        items:
          $ref: '#/definitions/report.OrdersByTimeOfDayResponse_Unit'
        type: array
    type: object
  report.OrdersByTimeOfDayResponse_Unit:
    properties:
      order_count:
        type: integer
      time_of_day:
        type: string
    type: object
  report.PurchasesByBrandsResponse:
    properties:
      units:
        items:
          $ref: '#/definitions/report.PurchasesByBrandsResponse_Unit'
        type: array
    type: object
  report.PurchasesByBrandsResponse_Unit:
    properties:
      brand:
        type: string
      purchase_count:
        type: integer
    type: object
info:
  contact: {}
  description: SSO service API.
  title: SSO API
  version: "1.0"
paths:
...
  /admin/report/average-processing-time/{period}:
    get:
      parameters:
      - description: Period (today, yesterday, week, month, year)
        enum:
        - today
        - yesterday
        - week
        - month
        - year
        in: path
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-report_AverageOrderProcessingTimeResponse'
      security:
      - BearerAuth: []
      summary: Get average order processing time
      tags:
      - admin-report
  /admin/report/orders-by-time/{period}:
    get:
      parameters:
      - description: Period (today, yesterday, week, month, year)
        enum:
        - today
        - yesterday
        - week
        - month
        - year
        in: path
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-report_OrdersByTimeOfDayResponse'
      security:
      - BearerAuth: []
      summary: Get amount of orders by time of day
      tags:
      - admin-report
  /admin/report/purchases-by-brands/{period}:
    get:
      parameters:
      - description: Period (today, yesterday, week, month, year)
        enum:
        - today
        - yesterday
        - week
        - month
        - year
        in: path
        name: period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-report_PurchasesByBrandsResponse'
      security:
      - BearerAuth: []
      summary: Get purchases by brands
      tags:
      - admin-report
  /auth/logIn:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_phenirain_sso_internal_dto_auth.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_auth.AuthResponse'
      summary: Login user
      tags:
      - auth
...
  /auth/signUp:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_phenirain_sso_internal_dto_auth.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_auth.AuthResponse'
      summary: Register user
      tags:
      - auth
...
  /manager/order:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-api_OrdersResponse'
      security:
      - BearerAuth: []
      summary: Get all orders
      tags:
      - manager
  /manager/order/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-api_ExtendedOrderResponse'
      security:
      - BearerAuth: []
      summary: Get order by id
      tags:
      - manager
  /manager/order/{id}/cancel:
    post:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-string'
      security:
      - BearerAuth: []
      summary: Cancel order
      tags:
      - manager
  /manager/order/give:
    post:
      consumes:
      - application/json
      parameters:
      - description: Paid order request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/manager.PaidOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_phenirain_sso_internal_dto_response.Response-string'
      security:
      - BearerAuth: []
      summary: Give order
      tags:
      - manager
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"